+++ src/gpu/vector4D.cu
@@ -9,7 +9,7 @@ namespace VRRT {
 
   __device__
   Vector3D Vector4D::projectTo3D() {
-    double invW = 1.0 / w;
+    float invW = 1.0 / w;
     return Vector3D(x * invW, y * invW, z * invW);
   }
